+++ stm32/src/Types/PinDef.cpp
@@ -7,7 +7,7 @@ bool PinDef::read() const {
     return digitalRead(PIN) == HIGH;
 }
 
-uint8_t PinDef::readAnalog() const {
+uint32_t PinDef::readAnalog() const {
     return analogRead(PIN);
 }
 
@@ -15,7 +15,7 @@ void PinDef::write(bool state) {
     digitalWrite(PIN, state ? HIGH : LOW);
 }
 
-void PinDef::writeAnalog(uint8_t state) {
+void PinDef::writeAnalog(uint32_t state) {
     analogWrite(PIN, state);
 }
